docs(colours): update colours docs and structure
