+++ projects/canopy/src/lib/docs/foundation/colours/colours.stories.mdx
@@ -1,0 +1,86 @@
+import { Meta, Story } from '@storybook/addon-docs';
+
+<Meta title="Foundation/Colours" parameters={{ previewTabs: { canvas: { hidden: true } } }} />
+
+# Colours
+
+<p class="standfirst">We are a bright and colourful brand, and don't shy away from it. Our palette draws its colours from our umbrella, and defines our core colours.</p>
+
+## Usage
+
+Import the css into the `angular.json` file of your application:
+
+~~~json
+"styles": [
+  "node_modules/@legal-and-general/canopy/canopy.css"
+],
+~~~
+
+## Accessibility
+
+It's important we use colours that provide an accessible experience to every user.
+
+This means that all colour combinations must adhere to WCAG 2.0 level AA standards, and have a contrast ratio of at least 4.5:1 for normal text and 3:1 for large text. We recommend using <a href="https://webaim.org/resources/contrastchecker/" target="_blank">Colour Contrast Checker</a>, but there are many similar tools available.
+
+Normal text is defined as 16px regular, while Large text is defined as 18.66px if bold or 24px regular.
+
+Custom gradients should not be used anywhere in your project.
+
+## Palette
+
+### Core
+
+Colour is an integral part of our identity and toolkit.
+
+Our designs will always use at least one of the core colours as the foundation for an application's colour palette.
+
+Primary colours are used for:
+
+- Key actions such as primary call to action buttons
+- Highlighted content
+- Form field elements
+
+
+
+<Story id="internal-colours--core-colours"></Story>
+
+
+
+### Secondary
+
+The rest of our colour palette expands on the four core colours. We use complementary lighter and darker colours to create variation and definition, particularly in:
+
+- Data visualisation
+- Illustration
+- Infographics
+- Brand icons
+
+
+
+<Story id="internal-colours--secondary-colours"></Story>
+
+
+
+### Tints
+
+For digital use only, we have also created a range of tints to be used within our UI elements.
+
+
+
+<Story id="internal-colours--tints-colours"></Story>
+
+
+
+### Greyscale
+
+Our black and grey palette is mostly reserved for typography, but we also use this range of tones for:
+
+- Iconography
+- Backgrounds
+- Lines and UI elements
+
+<Story id="internal-colours--greyscale-colours"></Story>
+
+## Further information
+
+For guidance on the use of colours for marketing and print, visit the <a href="https://www.legalandgeneralbrandhub.com/b/www/c/en-GB/BrandKnowledge/Guideline/2?Page=5" target="_blank">L&G Brand Hub</a>.
