--- projects/canopy/src/styles/color.notes.ts
@@ -1,12 +1,0 @@
-export const notes = `
-Provides available brand colours and tints via CSS variables
-
-## Usage
-Import the css into the angular.json file of your application, the variables should now be available for use.
-
-~~~json
-  "styles": [
-    "node_modules/@legal-and-general/canopy/canopy.css"
-  ],
-~~~
-`;
