+++ projects/canopy/src/styles/variables/_colors.scss
@@ -35,8 +35,8 @@
   --color-poppy-red-lightest: #ffeaec;
   --color-poppy-red-light: #f09491;
   --color-poppy-red: #e22922;
-
   --color-poppy-red-dark: #b50005;
+  --color-poppy-red-darkest: #671210;
 
   --color-dandelion-yellow-lightest: #fdf8cd;
   --color-dandelion-yellow-light: #ffef67;
